던전1, 2 추가
- 던전 기믹과 관련된 변경들
  - 발판, 디스펜서...

+++ Assets/Scripts/Enemy/ProjectileBullet.cs
@@ -4,7 +4,7 @@ using Photon.Pun;
 public class ProjectileBullet : MonoBehaviourPun
 {
     [SerializeField] private float speed = 20f;
-    [SerializeField] private float damage = 10f;
+    [SerializeField] private int damage = 10;
     [SerializeField] private float lifeTime = 3f;
 
     private void Start()
